Progress

+++ src/Zafiro.FileSystem/ISyncAction.cs
@@ -6,6 +6,6 @@ public interface ISyncAction
 {
     public IZafiroFile Source { get; set; }
     public IZafiroFile Destination { get; set; }
-    public IObservable<RelativeProgress<int>> Progress { get; }
+    public IObservable<RelativeProgress<long>> Progress { get; }
     public IObservable<Result> Sync();
 }
